% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energidataservice.R
\name{energidataservice}
\alias{energidataservice}
\title{Download data from Energidataservice.dk}
\usage{
energidataservice(
  dataset = "ProductionConsumptionSettlement",
  most.recent = NA,
  all = FALSE,
  start = lubridate::floor_date(Sys.time(), "year"),
  end = Sys.time()
)
}
\arguments{
\item{most.recent}{integer: No. of observations, starting with most recent}

\item{all}{bool: Used to get entire series. Defaults to FALSE.}

\item{start}{time of first point in data series. Format="YYYY-mm-dd" or "YYYY-mm-dd HH:MM". Defaults to Jan. 1st current year.}

\item{end}{time of last point in data series. Format="YYYY-mm-dd" or "YYYY-mm-dd HH:MM". Defaults to now.}
}
\value{
A data frame with production and consumption data
}
\description{
Downloads time-bound production, consumption and price data from
Danish TSO Energinet via its free API, found on api.energidataservice.dk
The API does not require user and/or token.
See energidataservice.dk for documentaion.
}
\details{
This function downloads settlement, i.e. validated data

Last updated 2022-12-17 by Magnus Gottlieb
}
